git-svn-id: https://svn.r-project.org/R/branches/R-BIGSEXP@68718 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/errors.c
@@ -80,7 +80,7 @@ static char * R_ConciseTraceback(SEXP call, int skip);
 
 static void reset_stack_limit(void *data)
 {
-    unsigned int *limit = (unsigned int *) data;
+    uintptr_t *limit = (uintptr_t *) data;
     R_CStackLimit = *limit;
 }
